requirements.txt
colorama==0.3.7
termcolor==1.1.0
PyYAML=3.12
